test4/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
requests>=2.31.0

